requirements.txt
asyncpg==0.20.1
fastapi==0.60.1
gunicorn==20.0.4
httpx==0.13.*
jsonschema==3.2.0
lru-dict==1.1.6
neo4j-driver==4.0.2
numpy==1.19.1
pydantic==1.6.1
git+https://github.com/ranking-agent/reasoner-pydantic#egg=reasoner-pydantic
pyyaml==5.3.1
git+https://github.com/ranking-agent/reasoner#egg=reasoner
redis==3.5.3
uvicorn==0.11.7